Fix bug when length is not positive
The length of the annotation has to be positive. Otherwise, it is not
displayed. I added a second error type to iTermAnnotationError. Because
invalidOptions is not very specific, I renamed it to
xOrYImpliesXYAndLength. The error messages were removed since the error
type names are already descriptive and they are difficult to handle.