Namespace, setCurrentDir and annotations in iTerm2 (#6)
I added a namespace for iTerm2 to ANSIEscapeCode. iTerm supports many
proprietary escape codes that I want to support in the future. An own
namespace makes it more clear which escape codes work only on iTerm.

Also, I rename setCwd to setCurrentDir because it matches the name of
the escape code in the docs of iTerm2 better.

I moved iTermAnnotationOptions to its own file because I do not like
too large files because I lose the overview. Also, the annotation
options do not have to do anything with the standard ANSI escape codes.

I added a few more examples for annotations to show how they work in
iTerm2. While doing so, I found a bug when the length is not positive.
The length of the annotation has to be positive. Otherwise, it is not
displayed. I added a second error type to iTermAnnotationError. Because
invalidOptions is not very specific, I renamed it to
xOrYImpliesXYAndLength. The error messages were removed since the error
type names are already descriptive and they are difficult to handle.

Also, I added many tests and comments for iTermAnnotationOptions,
setCurrentDir and annotations.

https://iterm2.com/documentation-escape-codes.html

See pull request https://github.com/ikelax/swift-ansi-escapes/pull/6