Remove isAppleTerminal and macros for os
I removed the macros that make code only available on specific OSes.
swift-dotenv uses Foundations' ProcessInfo.processInfo.environment on
Linux [0]. Additionally, this makes it possible to generate ANSI escape
codes for Windows while being on macOS, for example.

By setting the default value of terminal as the environment variable
TERM_PROGRAM, it is possible to test the method without having to mock
ProcessInfo.

[0]: https://github.com/thebarndog/swift-dotenv