Finish implementation of image (#9)
The height and the width have to be positive, since non-positive values
do not really make sense here. These values count the number of
character cells used to display the image.

The image was taken from
https://www.pexels.com/photo/people-along-narrow-alley-2181111/.

The resources used in the executable target "Image" and the test target
were declared explicitly such that the Swift compiler does not throw
warnings due to this.

ImageOptions was extracted to its own file because displaying images
seems to be a feature of iTerm2 and not part of the standard ANSI escape
codes.

`testCases` in `Annotation.swift` was declared as `private` because
otherwise, there is a name collision with `testCases` in `Image.swift`.

I renamed the test in `SetCurrentDir.swift` to `setCurrentDir` because
it is the name of the method tested.

Currently, the tests for `image` only run with `swift test [--filter
Image]`. XCode does not find the image due to reasons unknown to me. I
will investigate this later.

At the moment, the library does not support all options of the image
displaying feature in iTerm2. I will add the rest later.

See also: https://iterm2.com/documentation-images.html
See pull request: https://github.com/ikelax/swift-ansi-escapes/pull/9